arregla productos y posiblemente la base de datos

+++ src/main/java/eus/fpsanturtzilh/repositories/MaterialakRepository.java
@@ -8,7 +8,7 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MaterialakRepository extends JpaRepository<Materialak, Long> {
+public interface MaterialakRepository extends JpaRepository<Materialak, Integer> {
     @EntityGraph(attributePaths = {"maileguak"})
     List<Materialak> findAll();
